feat: Refactor ProtocolAdapter2 and remove ProtocolAdapter2Bridge for FSM-based redesign
