+++ src/main/java/com/hivemq/adapter/sdk/api/ProtocolAdapter2.java
@@ -16,43 +16,62 @@
 package com.hivemq.adapter.sdk.api;
 
 import com.hivemq.adapter.sdk.api.exceptions.ProtocolAdapterException;
+import com.hivemq.adapter.sdk.api.model.ProtocolAdapterStartInput;
+import com.hivemq.adapter.sdk.api.model.ProtocolAdapterStartOutput;
+import com.hivemq.adapter.sdk.api.model.ProtocolAdapterStopInput;
+import com.hivemq.adapter.sdk.api.model.ProtocolAdapterStopOutput;
+import com.hivemq.adapter.sdk.api.services.ModuleServices;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.atomic.AtomicReference;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * Simplified protocol adapter interface for the FSM-based redesign.
  * <p>
- * Key differences from {@link ProtocolAdapter}:
- * <ol>
- *   <li>{@link #connect} returns synchronously or throws — no CompletableFuture</li>
- *   <li>{@link #disconnect} returns synchronously or throws — no CompletableFuture</li>
- *   <li>Connection state is managed by the caller (the wrapper), not by the adapter itself</li>
- *   <li>Adapter should NOT manage its own state</li>
- * </ol>
- * <p>
- * Each adapter module provides its own implementation of this interface, typically by extending
- * {@link ProtocolAdapter2Bridge} and overriding {@link #supportsSouthbound()}.
+ * Implementations provide access to the wrapped legacy {@link ProtocolAdapter} and module services.
+ * The default methods implement the synchronous lifecycle expected by the FSM wrapper.
  */
 public interface ProtocolAdapter2 {
 
     /**
+     * Returns the underlying legacy protocol adapter instance.
+     *
+     * @return wrapped legacy adapter
+     */
+    @NotNull
+    ProtocolAdapter getLegacyAdapter();
+
+    /**
+     * Returns the module services used when delegating lifecycle calls to the legacy adapter.
+     *
+     * @return module services
+     */
+    @NotNull
+    ModuleServices getModuleServices();
+
+    /**
      * Get the adapter's unique identifier.
      */
-    @NotNull
-    String getId();
+    default @NotNull String getId() {
+        return getLegacyAdapter().getId();
+    }
 
     /**
-     * Get adapter information (protocol type, capabilities, etc.)
+     * Get adapter information (protocol type, capabilities, etc.).
      */
-    @NotNull
-    ProtocolAdapterInformation getProtocolAdapterInformation();
+    default @NotNull ProtocolAdapterInformation getProtocolAdapterInformation() {
+        return getLegacyAdapter().getProtocolAdapterInformation();
+    }
 
     /**
-     * Check if this adapter supports Southbound (MQTT to Device) communication.
+     * Check if this adapter supports southbound (MQTT to device) communication.
      *
-     * @return true if southbound is supported, false if this is a read-only adapter
+     * @return true if southbound is supported
      */
     default boolean supportsSouthbound() {
-        return false;
+        return getLegacyAdapter().getProtocolAdapterInformation().getCapabilities().contains(ProtocolAdapterCapability.WRITE);
     }
 
     /**
@@ -61,44 +80,84 @@ public interface ProtocolAdapter2 {
      *
      * @throws ProtocolAdapterException if configuration is invalid
      */
-    void precheck() throws ProtocolAdapterException;
+    default void precheck() throws ProtocolAdapterException {
+        // Legacy adapters do not expose a precheck API.
+    }
 
     /**
      * Establish connection to the device/service.
-     * Called for BOTH northbound and southbound connections.
-     * <p>
-     * This method should:
-     * <ul>
-     *   <li>Establish the physical/logical connection</li>
-     *   <li>Validate connectivity (handshake, auth, etc.)</li>
-     *   <li>Return when connection is ready</li>
-     *   <li>Throw on any failure</li>
-     * </ul>
      *
      * @param direction the connection direction (northbound vs southbound)
      * @throws ProtocolAdapterException on connection failure
      */
-    void connect(@NotNull ProtocolAdapterConnectionDirection direction) throws ProtocolAdapterException;
+    default void connect(final @NotNull ProtocolAdapterConnectionDirection direction) throws ProtocolAdapterException {
+        if (direction.isSouthbound()) {
+            // Legacy adapters handle southbound internally during start.
+            return;
+        }
+        final ProtocolAdapterStartInput input = () -> getModuleServices();
+        final CompletableFuture<Void> startFuture = new CompletableFuture<>();
+        final AtomicReference<String> errorMessage = new AtomicReference<>();
+        getLegacyAdapter().start(input, new ProtocolAdapterStartOutput() {
+            @Override
+            public void startedSuccessfully() {
+                startFuture.complete(null);
+            }
+
+            @Override
+            public void failStart(final @NotNull Throwable throwable, final @Nullable String message) {
+                errorMessage.set(message);
+                startFuture.completeExceptionally(throwable);
+            }
+        });
+        try {
+            startFuture.get();
+        } catch (final InterruptedException e) {
+            Thread.currentThread().interrupt();
+            throw new ProtocolAdapterException("Adapter start interrupted for '" + getId() + "'", e);
+        } catch (final ExecutionException e) {
+            final Throwable cause = e.getCause() != null ? e.getCause() : e;
+            throw new ProtocolAdapterException(
+                    "Adapter start failed for '" + getId() + "': " + errorMessage.get(), cause);
+        }
+    }
 
     /**
      * Disconnect from the device/service.
-     * <p>
-     * This method should:
-     * <ul>
-     *   <li>Gracefully close the connection</li>
-     *   <li>Release resources</li>
-     *   <li>Return when cleanup is complete</li>
-     *   <li>NOT throw on failure (log errors instead)</li>
-     * </ul>
      *
      * @param direction the connection direction (northbound vs southbound)
      */
-    void disconnect(@NotNull ProtocolAdapterConnectionDirection direction);
+    default void disconnect(final @NotNull ProtocolAdapterConnectionDirection direction) {
+        if (direction.isSouthbound()) {
+            // Legacy adapters handle southbound internally during stop.
+            return;
+        }
+        final ProtocolAdapterStopInput input = new ProtocolAdapterStopInput() {};
+        final CompletableFuture<Void> stopFuture = new CompletableFuture<>();
+        getLegacyAdapter().stop(input, new ProtocolAdapterStopOutput() {
+            @Override
+            public void stoppedSuccessfully() {
+                stopFuture.complete(null);
+            }
+
+            @Override
+            public void failStop(final @NotNull Throwable throwable, final @Nullable String errorMessage) {
+                stopFuture.completeExceptionally(throwable);
+            }
+        });
+        try {
+            stopFuture.get();
+        } catch (final InterruptedException e) {
+            Thread.currentThread().interrupt();
+        } catch (final ExecutionException e) {
+            // Disconnect should not throw; caller handles error reporting.
+        }
+    }
 
     /**
      * Destroy the adapter instance.
-     * Called after both connections are disconnected.
-     * Release ALL resources including configuration.
      */
-    void destroy();
+    default void destroy() {
+        getLegacyAdapter().destroy();
+    }
 }
