+++ src/main/java/com/hivemq/adapter/sdk/api/factories/ProtocolAdapterFactory.java
@@ -18,28 +18,26 @@ package com.hivemq.adapter.sdk.api.factories;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hivemq.adapter.sdk.api.ProtocolAdapter;
 import com.hivemq.adapter.sdk.api.ProtocolAdapter2;
-import com.hivemq.adapter.sdk.api.ProtocolAdapter2Bridge;
 import com.hivemq.adapter.sdk.api.ProtocolAdapterInformation;
 import com.hivemq.adapter.sdk.api.config.ProtocolSpecificAdapterConfig;
 import com.hivemq.adapter.sdk.api.model.ProtocolAdapterInput;
 import com.hivemq.adapter.sdk.api.services.ModuleServices;
 import com.hivemq.adapter.sdk.api.tag.Tag;
-import org.jetbrains.annotations.NotNull;
-
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+import org.jetbrains.annotations.NotNull;
 
 /**
  * The factory is responsible for constructing and managing the lifecycle of the various aspects of the adapter
- * sub-systems. We bind this to the configuration types to we can provide tightly coupled implementation instances
+ * sub-systems. We bind this to the configuration types so we can provide tightly coupled implementation instances
  * responsible for adapter management.
  */
 public interface ProtocolAdapterFactory<E extends ProtocolSpecificAdapterConfig> {
 
     /**
-     * Returns Metadata related to the protocol adapter instance, including descriptions, iconography, categorisation et
-     * al
+     * Returns metadata related to the protocol adapter instance, including descriptions, iconography, categorisation,
+     * etc.
      *
      * @return the instance that provides the adapter information
      */
@@ -47,10 +45,10 @@ public interface ProtocolAdapterFactory<E extends ProtocolSpecificAdapterConfig>
     ProtocolAdapterInformation getInformation();
 
     /**
-     * This method is called by HiveMQ Edge to create an instance of the adapter
+     * This method is called by HiveMQ Edge to create an instance of the adapter.
      *
      * @param adapterInformation the information on the adapter
-     * @param input              wrapper object for various information for the adapter.
+     * @param input              wrapper object for various information for the adapter
      * @return the protocol adapter instance
      */
     @NotNull
@@ -87,7 +85,7 @@ public interface ProtocolAdapterFactory<E extends ProtocolSpecificAdapterConfig>
 
     /**
      * @param objectMapper the object mapper that converts the map to the actual tag
-     * @param tag      a map that is a tag
+     * @param tag          a map that is a tag
      * @return a parsed tag object for this adapter
      */
     default @NotNull Tag convertTagDefinitionObject(
@@ -106,21 +104,28 @@ public interface ProtocolAdapterFactory<E extends ProtocolSpecificAdapterConfig>
     }
 
     /**
-     * Creates a {@link ProtocolAdapter2} instance wrapping the given old-SDK adapter.
+     * Creates a {@link ProtocolAdapter2} instance wrapping the given legacy adapter.
      * <p>
-     * Override this method to return a module-specific {@link ProtocolAdapter2} implementation
-     * that encodes adapter-type-specific behavior (e.g., whether southbound is supported).
-     * <p>
-     * The default implementation returns a generic {@link ProtocolAdapter2Bridge} that determines
-     * southbound support from the adapter's {@link com.hivemq.adapter.sdk.api.ProtocolAdapterCapability#WRITE}
-     * capability.
+     * Module factories can override this method to return module-specific implementations.
+     * The default implementation returns an inline adapter that delegates to the legacy adapter through
+     * {@link ProtocolAdapter2}'s default methods.
      *
-     * @param protocolAdapter the old-SDK adapter instance created by {@link #createAdapter}
+     * @param protocolAdapter the legacy adapter instance created by {@link #createAdapter}
      * @param moduleServices  the per-module services for the adapter
      * @return the {@link ProtocolAdapter2} implementation for this adapter type
      */
     default @NotNull ProtocolAdapter2 createProtocolAdapter2(
             final @NotNull ProtocolAdapter protocolAdapter, final @NotNull ModuleServices moduleServices) {
-        return new ProtocolAdapter2Bridge(protocolAdapter, moduleServices);
+        return new ProtocolAdapter2() {
+            @Override
+            public @NotNull ProtocolAdapter getLegacyAdapter() {
+                return protocolAdapter;
+            }
+
+            @Override
+            public @NotNull ModuleServices getModuleServices() {
+                return moduleServices;
+            }
+        };
     }
 }
